Split out calculation
